BDOG-516 refactor out header utils
