+++ src/main/scala/uk/gov/hmrc/SbtAutoBuildPlugin.scala
@@ -22,10 +22,6 @@ import de.heikoseeberger.sbtheader.HeaderPlugin.autoImport._
 import de.heikoseeberger.sbtheader.{AutomateHeaderPlugin, CommentStyle, FileType}
 import sbt.Keys._
 import sbt.{Setting, _}
-import uk.gov.hmrc.RepositoryYamlUtils.{Private, Public, RepoVisibility}
-
-import scala.io.Source
-import scala.util.Try
 
 object SbtAutoBuildPlugin extends AutoPlugin {
 
@@ -96,15 +92,9 @@ object PublishSettings {
 }
 
 // Enforce a standard licence header across all HMRC
+// public repo -> apache v2
+// private repo -> only copyright headers
 object HeaderSettings {
-  import Extensions.RichTry
-
-  val license = new File("LICENSE")
-  val repositoryYamlFile = new File("repository.yaml")
-
-  val expectedLicenceText =
-    """Apache License
-      |                           Version 2.0, January 2004""".stripMargin
 
   val commentStyles: Map[FileType, CommentStyle] = Map(
     FileType.scala -> CommentStyle.cStyleBlockComment,
@@ -112,56 +102,10 @@ object HeaderSettings {
     FileType("html") -> CommentStyle.twirlStyleBlockComment
   )
 
-  def shouldGenerateHeaders(force: Boolean): Boolean = {
-
-    // .right projection required to remain backwards compatible with scala 2.10 cross build (for sbt 0.13)
-    lazy val errorOrVisibility = for {
-      yamlString <- RepositoryYamlUtils.loadRepositoryYamlFile(new File("")).right
-      yaml <- RepositoryYamlUtils.loadYaml(yamlString).right
-      repoVisibility <- RepositoryYamlUtils.getRepoVisiblity(yaml).right
-      _ <- checkLicenceFile(repoVisibility).right
-    } yield repoVisibility
-
-    def checkLicenceFile(repoVisibility: RepoVisibility): Either[String, Unit] = repoVisibility match {
-      case Private =>
-        if(license.exists()) Left(s"LICENSE file exists but the repository is marked as private. Please remove it")
-        else Right(())
-      case Public =>
-        if (license.exists()) {
-          Try {
-            val f = Source.fromFile(license)
-            val content = f.mkString
-            f.close()
-            content
-          }.toEither.left.map(_ => s"Problem reading LICENSE file")
-              // .right projection required to remain backwards compatible with scala 2.10 cross build (for sbt 0.13)
-              .right.flatMap(c =>
-                  if (c.contains(expectedLicenceText)) Right(())
-                  else Left(s"The LICENSE file does not contain the appropriate Apache V2 licence")
-          )
-        }
-        else Left(s"No LICENSE file exists but the repository is marked as public")
-    }
-
-    if (force) {
-      SbtAutoBuildPlugin.logger.info(s"SbtAutoBuildPlugin - ${SbtAutoBuildPlugin.forceLicenceHeader.key.label} setting was set to true, Apache 2.0 licence file headers will be generated regardless")
-      true
-    } else errorOrVisibility match {
-      case Right(Public) =>
-        SbtAutoBuildPlugin.logger.info("SbtAutoBuildPlugin - repository is marked public. Licence headers will be added to all source files")
-        true
-      case Right(Private) =>
-        SbtAutoBuildPlugin.logger.info("SbtAutoBuildPlugin - repository is marked private. No licence headers will be added to source files, only a copyright notice")
-        false
-      case Left(error) =>
-        sys.error(s"SbtAutoBuildPlugin - Error, please fix: $error")
-    }
-  }
-
   def apply(forceSourceHeader: SettingKey[Boolean]): Seq[Setting[_]] = {
     Seq(
       headerLicense := {
-        if (shouldGenerateHeaders(forceSourceHeader.value))
+        if (HeaderUtils.shouldGenerateHeaders(forceSourceHeader.value))
           Some(HeaderLicense.ALv2(SbtAutoBuildPlugin.currentYear, "HM Revenue & Customs"))
         else Some(HeaderLicense.Custom(
           s"""|Copyright ${SbtAutoBuildPlugin.currentYear} HM Revenue & Customs
